fix compiler warnings

+++ arm9/include/core.h
@@ -34,14 +34,14 @@ USA
 #include <string.h>
 #include <stdlib.h>
 
-#define MAP_RELOAD      0x80000000
-#define MAP_PPU         0x81000000
-#define MAP_CPU         0x82000000
-#define MAP_DSP         0x83000000
-#define MAP_LOROM_SRAM  0x84000000
-#define MAP_HIROM_SRAM  0x85000000
-#define MAP_NONE        0x86000000
-#define MAP_LAST        0x8F000000
+#define MAP_RELOAD      (u32)(0x80000000)
+#define MAP_PPU         (u32)(0x81000000)
+#define MAP_CPU         (u32)(0x82000000)
+#define MAP_DSP         (u32)(0x83000000)
+#define MAP_LOROM_SRAM  (u32)(0x84000000)
+#define MAP_HIROM_SRAM  (u32)(0x85000000)
+#define MAP_NONE        (u32)(0x86000000)
+#define MAP_LAST        (u32)(0x8F000000)
 
 #define bzero(p, s)	memset(p, 0, s)
